% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Essentials.R
\name{buildpath}
\alias{buildpath}
\title{Recursively generate Directories}
\usage{
buildpath(dbjoin, dir = "", foldersetup = F)
}
\arguments{
\item{dbjoin}{Tibble; Tibble containing the joined ARTMO Database}

\item{foldersetup}{boolean; Do you want to automatically create a folder structure based on the specified directory?}

\item{mydir}{character; Location of a Directory}
}
\description{
After the access, cleaning and join of tables they can be exported from R. 
The *buildpath* function is useful to create the necessary folder infrastructure on the file system based on a specified directory
}
