% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYSQL_Essential.R
\name{connect.artmo}
\alias{connect.artmo}
\title{Connect to ARTMO}
\usage{
connect.artmo(user, password, host, database = NULL, getDB = F)
}
\arguments{
\item{user}{character; MYSQL Username (default = root)}

\item{password}{character; MYSQL password (default = 123456)}

\item{host}{character; Host of the MySQL Server (default = localhost)}

\item{database}{character; Name of the Database you would like to use. If left empty all the databases are returned.}
}
\description{
This function provides the core connection to the MySQL backend. 
By digitizing the MySQL credential and optionally the database R automatically connects to the MYSQL backend.
}
