% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYSQL_ColumnRenaming.R
\name{cost.tableclass}
\alias{cost.tableclass}
\title{Functions for renaming the columns in the Scripts from Juan Pablo}
\usage{
cost.tableclass(x)
}
\description{
Convert any Matlab matrix to the right tibble format in R
Formats everything in tidy format
}
