% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYSQL_Essential.R
\name{db.master}
\alias{db.master}
\title{Get the ARTMO Master Table}
\usage{
db.master(con)
}
\arguments{
\item{con}{MariaDBConnection; connection to a ARTMO Database}
}
\description{
This function returns the master table. 
This functionality is limited to the LUT since the other apporaches do not provide a master table.
}
