% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYSQL_Essential.R
\name{db.sensor}
\alias{db.sensor}
\title{Get the ARTMO Sensor Information}
\usage{
db.sensor(con)
}
\arguments{
\item{con}{MariaDBConnection; connection to a ARTMO Database}
}
\description{
This function returns the sensor information by iteration. 
This function has not yet been tested for each of each approach.
}
