% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYSQL_Essential.R
\name{db.tables}
\alias{db.tables}
\title{List ARTMO Tables}
\usage{
db.tables(con)
}
\arguments{
\item{con}{MariaDBConnection; connection to a ARTMO Database}
}
\description{
This function lists all the tables available in the ARTMO MYSQL backend. 
It helps to understand how much data is stored by table and to specifically address 
the table on interest during the concatentation of tables
}
