% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYSQL_Joiner.R
\name{doJoin}
\alias{doJoin}
\title{Join ARTMO Tables}
\usage{
doJoin(tabs, removeid = F)
}
\arguments{
\item{tabs}{nested tibble; table extraction done by `getTabs`}

\item{removeid}{boolean; Shall the ID be removed?}
}
\description{
This is the third function necessary for the combination of the extracted Tables
Here all the Tables generated beforehand are combined in the respective order with the others.
The tables are saved as tibbles and differred by the Table Type (MLRA, LUT CF, VIs).
Aside the **con** parameter the function needs the linkage inormation provided by the `getTabs`
}
