% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Essentials.R
\name{formatARTMO}
\alias{formatARTMO}
\title{Condense ARTMO Dataset}
\usage{
formatARTMO(jtable, standard = T, model = NULL, parameters = NULL,
  statistics = NULL)
}
\arguments{
\item{jtable}{Tibble; Tibble containing the joined ARTMO Database}

\item{standard}{boolean; Standard way of standardization (model, parameter and statistics are ignored)}

\item{model}{character; Reduction by Model (Optional)}

\item{parameters}{character;  Reduction by Parameter(s) (Optional)}

\item{statistics}{character;  Reduction by certain statistics (Optional)}
}
\description{
This function helps the user to format the Results table based. 
This functions allows to reduce the information of a table on the level of statistical accuracy
}
