% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Essentials.R
\name{formatTidy}
\alias{formatTidy}
\title{Create Tidy Data Arrays from Results}
\usage{
formatTidy(jtable)
}
\arguments{
\item{jtable}{Tibble; Tibble containing the joined ARTMO Database}
}
\description{
This function helps the user to format the Results table based. 
This functions allows to reduce the information of a table on the level of statistical accuracy
}
