% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYSQL_Joiner.R
\name{getLinks}
\alias{getLinks}
\title{Return ARTMO Table Relationships}
\usage{
getLinks(con)
}
\arguments{
\item{con}{MariaDBConnection; connection to a ARTMO Database}
}
\description{
This is the first function necessary for the combination of the MYSQL back.end data.
It searches for the links between the single tables stored in MYSQL. 
Thereby a tibble is generated containing the name of the databse, Table, its ID as well as the Type of the Table (MLRA, LUT etc.).
The dataframe is added by an information of the rank (position in table sequence) as well as whether it is a metadata table or a results table.
This dataframe is used for the `getTabs` function.
}
