% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYSQL_Joiner.R
\name{getMYSQL}
\alias{getMYSQL}
\title{MYSQL Processor}
\usage{
getMYSQL(con, removeid = F)
}
\arguments{
\item{con}{MariaDBConnection; connection to a ARTMO Database}

\item{removeid}{boolean; Shall the ID be removed?}
}
\description{
This function unites all steps individually done for accessing the MYSQL back-end.
This allows the combined action of the three functions `getLinks`,`getTabs` and `doJoin`
}
