% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYSQL_Joiner.R
\name{getTabs}
\alias{getTabs}
\title{Return ARTMO Tables}
\usage{
getTabs(con, links)
}
\arguments{
\item{con}{MariaDBConnection; connection to a ARTMO Database}

\item{links}{tibble; Link between the MySQL tables}
}
\description{
This is the second function necessary for the extraction and transformation of the MYSQL back-end data.
The function is necessary for reading the MYSQL Tables based on their link to each other.
Aside the **con** parameter the function needs the linkage inormation provided by the `getLinks`.
}
