% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Essentials.R
\name{rawTrans}
\alias{rawTrans}
\title{ARTMO Blob Converter}
\usage{
rawTrans(input)
}
\arguments{
\item{input}{raw; Raw Binary as stored in the MYSQL ARTMO Backend}
}
\description{
The ARTMO Data mostly lies within MYSQL Tables. MySQL has the possibility to store data both in so called "Blobs".
Some of the Blobs the ARTMO Software generates are stored in Binaric within the Matlab Files. In order to address these properly,
we created this function. It searches the different parts of the Binary File including flags for the datatypes, length and others.
}
