% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{spatDir}
\alias{spatDir}
\title{Link the Spatial Directories to the Databases}
\usage{
spatDir(table, dir = "", addraster = F)
}
\arguments{
\item{table}{Tibble; Tibble retrieved with the MYSQL extraction `getMYSQL`}

\item{dir}{character; Directory od the Datastructure built by `buildpath`}

\item{addraster}{boolean; Have the Rasters already been added and do you want to include them in the list?}
}
\description{
one
}
