% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{spatRes}
\alias{spatRes}
\title{Add the Spaial Results to the Dataframe}
\usage{
spatRes(spatial.df, col = "")
}
\arguments{
\item{spatial.df}{tibble; The spatial results will be appended to this tibble or data frame}

\item{col}{character; Column indluding the measurements}
}
\description{
This functions adds the Rasterfiles to the data frame corresponding to the Spatial
}
