% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.R
\name{spatRes}
\alias{spatRes}
\title{Add the Spaial Results to the Dataframe}
\usage{
spatRes(spatial.df, measuredCol = "")
}
\arguments{
\item{table}{Tibble; Tibble retrieved with the MYSQL extraction `getMYSQL`}

\item{dir}{character; Directory od the Datastructure built by `buildpath`}

\item{addraster}{boolean; Have the Rasters already been added and do you want to include them in the list?}
}
\description{
two
}
